requirements.txt
pre-commit
attrs
tqdm
dateparser
colorama
sql-metadata
openai==0.28.0
tiktoken
guidance==0.0.64
pyparsing==3.1.1
fiscalyear
pandas
sqlglot
tabulate
pytest
typeguard
rapidfuzz
records
SQLAlchemy
recognizers-text
recognizers-text-suite
emoji==1.7.0
azure-identity
transformers
nbformat
datasets==2.16.1
nltk
sentence-transformers
wikiextractor

rouge_score
git+https://github.com/google-research/bleurt.git